Add models JSDoc and rename token service JC-673
