+++ apps/angular/src/app/features/registration/registration-form/registration-form.component.ts
@@ -5,7 +5,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
 import { AuthorizationService } from '@js-camp/angular/core/services/authorization.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { ServerError } from '@js-camp/core/models/server-error';
+import { ServerError } from '@js-camp/core/models/server-error.model';
 import { FormValidationService } from '@js-camp/angular/core/services/form-validation.service';
 import { Observable, tap, ReplaySubject } from 'rxjs';
 import { CommonModule } from '@angular/common';
