+++ libs/core/models/server-error.model.ts
@@ -6,6 +6,6 @@ export type ServerError = {
 	/** Controls name. */
 	readonly controlName: string;
 
-	/** . */
+	/** An array of controls error messages. */
 	readonly controlErrors: ErrorMessage[];
 };
