--- libs/core/utils/enums/model-registration.enum.ts
@@ -1,9 +1,0 @@
-/** 1. */
-export enum RegistrationModel {
-	FirstName = 'firstName',
-	LastName = 'lastName',
-	Email = 'email',
-	Password = 'password',
-	ConfirmPassword = 'confirmPassword',
-	Avatar = 'avatar',
-}
